Created a super class for the current base menu for not having navigation buttons. It will mostly be used for the main menu but also it's nice to have a less cluttered navigation menu
